users/graphql/mutations/register_user.py

import graphene
from django.contrib.auth.models import User
from graphql_jwt.shortcuts import get_token, create_refresh_token
from ..types import UserType

class RegisterUser(graphene.Mutation):
    user = graphene.Field(UserType)
    token = graphene.String()
    refresh_token = graphene.String()

    class Arguments:
        username = graphene.String(required=True)
        password = graphene.String(required=True)
        email = graphene.String(required=True)

    def mutate(self, info, username, password, email):
        user = User.objects.create_user(username=username, email=email, password=password)
        token = get_token(user)
        refresh_token = create_refresh_token(user)
        return RegisterUser(user=user, token=token, refresh_token=refresh_token)
